Web+DB/requirements.txt
codegen==1.0
Django==2.1.2
numpy==1.16.4
opencv-python==4.1.0.25
patterns==0.3
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2018.6
view==0.1
